refactor: generalize bar chart
